GitHub CI [Mac]: don't require all non-free SDKs
Do not require all of the non-free SDKs to be present. User may opt in
only for one (and doesn't own the other, either).

+++ .github/scripts/macOS/prepare_nonfree_sdks.sh
@@ -4,24 +4,27 @@ cd /tmp
 
 # AJA
 AJA_DIRECTORY=$SDK_NONFREE_PATH/ntv2sdk
-unzip $SDK_NONFREE_PATH/ntv2sdkmac.zip
-mv /tmp/ntv2sdk* $AJA_DIRECTORY
-cd $AJA_DIRECTORY/ajalibraries/ajantv2/build
-xcodebuild -project ajantv2.xcodeproj
+if [ -f $SDK_NONFREE_PATH/ntv2sdkmac.zip ]; then
+        mv /tmp/ntv2sdk* $AJA_DIRECTORY
+        cd $AJA_DIRECTORY/ajalibraries/ajantv2/build
+        xcodebuild -project ajantv2.xcodeproj
+fi
 
 # DELTACAST
 DELTA_CACHE_INST=$SDK_NONFREE_PATH/VideoMasterHD_inst
-unzip /var/tmp/sdks-nonfree/VideoMaster_SDK_MacOSX.zip
-sudo installer -pkg VideoMaster_SDK.pkg -target /
-cd /Library/Frameworks
-sudo install_name_tool -change /Library/Frameworks/VideoMasterHD.framework/Versions/A/VideoMasterHD @executable_path/../Frameworks/VideoMasterHD.framework/Versions/A/VideoMasterHD /Library/Frameworks/VideoMasterHD.framework/VideoMasterHD
-sudo install_name_tool -id @executable_path/../Frameworks/VideoMasterHD.framework/Versions/A/VideoMasterHD /Library/Frameworks/VideoMasterHD.framework/VideoMasterHD
-sudo install_name_tool -change /Library/Frameworks/libVideoMasterHD_SP.dylib @executable_path/../Frameworks/libVideoMasterHD_SP.dylib /Library/Frameworks/VideoMasterHD.framework/VideoMasterHD
-sudo install_name_tool -id @executable_path/../Frameworks/libVideoMasterHD_SP.dylib /Library/Frameworks/libVideoMasterHD_SP.dylib
-sudo install_name_tool -id @executable_path/../Frameworks/VideoMasterHD_Audio.framework/Versions/A/VideoMasterHD_Audio /Library/Frameworks/VideoMasterHD_Audio.framework/Versions/A/VideoMasterHD_Audio
-sudo install_name_tool -change /Library/Frameworks/VideoMasterHD.framework/Versions/A/VideoMasterHD @executable_path/../Frameworks/VideoMasterHD.framework/Versions/A/VideoMasterHD /Library/Frameworks/VideoMasterHD_Audio.framework/Versions/A/VideoMasterHD_Audio
-#sudo cp -a VideoMasterHD.framework VideoMasterHD_Audio.framework libVideoMasterHD_SP.dylib $(xcrun --show-sdk-path)/System/Library/Frameworks
-mkdir $DELTA_CACHE_INST
-sudo cp -a VideoMasterHD.framework VideoMasterHD_Audio.framework libVideoMasterHD_SP.dylib $DELTA_CACHE_INST
-cd -
-sudo rm -rf /Library/Frameworks/VideoMasterHD* # ensure that only the copy above is used
+if [ -f $SDK_NONFREE_PATH/VideoMaster_SDK_MacOSX.zip ]; then
+        unzip $SDK_NONFREE_PATH/VideoMaster_SDK_MacOSX.zip
+        sudo installer -pkg VideoMaster_SDK.pkg -target /
+        cd /Library/Frameworks
+        sudo install_name_tool -change /Library/Frameworks/VideoMasterHD.framework/Versions/A/VideoMasterHD @executable_path/../Frameworks/VideoMasterHD.framework/Versions/A/VideoMasterHD /Library/Frameworks/VideoMasterHD.framework/VideoMasterHD
+        sudo install_name_tool -id @executable_path/../Frameworks/VideoMasterHD.framework/Versions/A/VideoMasterHD /Library/Frameworks/VideoMasterHD.framework/VideoMasterHD
+        sudo install_name_tool -change /Library/Frameworks/libVideoMasterHD_SP.dylib @executable_path/../Frameworks/libVideoMasterHD_SP.dylib /Library/Frameworks/VideoMasterHD.framework/VideoMasterHD
+        sudo install_name_tool -id @executable_path/../Frameworks/libVideoMasterHD_SP.dylib /Library/Frameworks/libVideoMasterHD_SP.dylib
+        sudo install_name_tool -id @executable_path/../Frameworks/VideoMasterHD_Audio.framework/Versions/A/VideoMasterHD_Audio /Library/Frameworks/VideoMasterHD_Audio.framework/Versions/A/VideoMasterHD_Audio
+        sudo install_name_tool -change /Library/Frameworks/VideoMasterHD.framework/Versions/A/VideoMasterHD @executable_path/../Frameworks/VideoMasterHD.framework/Versions/A/VideoMasterHD /Library/Frameworks/VideoMasterHD_Audio.framework/Versions/A/VideoMasterHD_Audio
+        #sudo cp -a VideoMasterHD.framework VideoMasterHD_Audio.framework libVideoMasterHD_SP.dylib $(xcrun --show-sdk-path)/System/Library/Frameworks
+        mkdir $DELTA_CACHE_INST
+        sudo cp -a VideoMasterHD.framework VideoMasterHD_Audio.framework libVideoMasterHD_SP.dylib $DELTA_CACHE_INST
+        cd -
+        sudo rm -rf /Library/Frameworks/VideoMasterHD* # ensure that only the copy above is used
+fi
